+++ CHANGELOG.md
@@ -5,6 +5,9 @@ underlying symbol, show the `defalias` call in the source code, don't
 show duplicates in the Aliases list, and cross-reference the Elisp
 manual.
 
+Fixed an issue where keymaps with prefix keys were not rendered
+correctly.
+
 Fixed an issue when looking at variable docs where we would visit the
 definition buffer without the user requesting it.